Module figtree::types

A collection of types that define a figtree document.

These types are re-exported in the main module because there aren't too many of them, and because they're useful when testing equality or building figtree documents from scratch. The types module is also made available to allow explicit namespaced imports of these types.

Examples

let mut doc = Document::new();
let mut node = doc.new_node_or_get("node_name");
node.insert_attr(
    "key".to_string(),
    Value::new_int(4032));

Structs

Document
A struct representing a parsed figtree document.

Node
A struct representing an individual node in a parsed document

Enums

Value
A type to represent a figtree value

Type Definitions

Dict
A type to represent a figtree dict

List
A type to represent a figtree list