Enum figtree::types::Value

pub enum Value {
    Str(String),
    Int(i64),
    Float(f64),
    Bool(bool),
    Ident(String),
    Dict(Dict),
    List(List),
    Null,
}

A type to represent a figtree value

Generally, this is obtained by getting the value of an attribute on a Node, although there are also methods to construct all the kinds of this type.

Examples

let value = Value::new_string("hello!");
assert!(value.get_str() == Some("hello!"));
assert!(value.get_int() == None);

Variants

Str

Int

Float

Bool

Ident

Dict

List

Null

Methods

impl Value

fn new_string<S>(s: S) -> Self where S: Into<String>

Construct a new string Value.

fn new_ident<S>(s: S) -> Self where S: Into<String>

Construct a new identifier Value.

fn new_int(s: i64) -> Self

Construct a new integer Value.

fn new_float(s: f64) -> Self

Construct a new float Value.

fn new_bool(s: bool) -> Self

Construct a new boolean Value.

fn new_null() -> Self

Construct a new null Value.

fn from_parsed_value(val: ParsedValue) -> Self

fn get_str(&self) -> Option<&str>

Extract the contained value if it is a string.

fn get_int(&self) -> Option<i64>

Extract the contained value if it is an integer

fn get_float(&self) -> Option<f64>

Extract the contained value if it is a float

fn get_bool(&self) -> Option<bool>

Extract the contained value if it is a boolean

fn get_ident(&self) -> Option<&str>

Extract the contained (&str) value if it is an identifier

fn get_dict(&self) -> Option<&Dict>

Extract the contained value if it is a dict

fn get_list(&self) -> Option<&[Value]>

Extract the contained value as a slice if it is a list

fn is_null(&self) -> bool