Struct figtree::types::Node

pub struct Node {
    // some fields omitted
}

A struct representing an individual node in a parsed document

Examples

Manipulating subnodes:

use figtree::types::*;
let mut node = Node::new();
assert!(node.is_empty());
assert!(!node.has_nodes());

node.new_node_or_get("subnode");
assert!(node.node_count() == 1);

{ // appease the borrow checker
    let subnode = node.get_node("subnode").expect("no such node");
    assert!(subnode == &Node::new());  // creates a blank subnode
}

{ // appease the borrow checker
    let mut subnode = node.get_node_mut("subnode").expect("no such node");
    let sub_subnode = subnode.new_node_or_get("sub_subnode");
    // etc.
}

Manipulating attributes

use figtree::types::*;
let mut node = Node::new();
assert!(node.is_empty());
assert!(!node.has_attrs());

node.insert_attr("key", Value::new_int(5));
assert!(node.attr_count() == 1);

Methods

impl Node

fn new() -> Self

Construct a new, empty node

fn new_node_or_get<S>(&mut self, name: S) -> &mut Self where S: Into<String>

Construct a new node and automatically insert it as a subnode.

Returns a mutable reference to the new node. If there is a subnode already present with the given name, this method will not insert a new node and instead just return the old node.

fn insert_node<S>(&mut self, name: S, node: Node) -> Option<Node> where S: Into<String>

Inserts a node into this node as a subnode.

If there is already a node with the given name, replace it and return the old node.

fn delete_node<S>(&mut self, name: S) -> Option<Node> where S: Into<String>

Remove a subnode from this node.

Returns the deleted node.

fn get_node<S>(&self, name: S) -> Option<&Self> where S: Into<String>

Get a reference to the specified subnode

fn get_node_mut<S>(&mut self, name: S) -> Option<&mut Self> where S: Into<String>

Get a mutable reference to the specified subnode

fn iter_nodes(&self) -> Iter<String, Node>

Get an iterable of (&name, &node) pairs

fn insert_attr<S>(&mut self, name: S, value: Value) -> Option<Value> where S: Into<String>

Insert a new value into this node.

If there is already a value with the given name, replace it and return the old value.

fn delete_attr<S>(&mut self, name: S) -> Option<Value> where S: Into<String>

Remove an attribute from this node.

Returns the deleted value.

fn get_attr<S>(&self, name: S) -> Option<&Value> where S: Into<String>

Get a reference to the specified attribute value

fn get_attr_mut<S>(&mut self, name: S) -> Option<&mut Value> where S: Into<String>

Get a mutable reference to the specified attribute value

fn iter_attrs(&self) -> Iter<String, Value>

Get an iterable of (&name, &attribute) pairs

fn is_empty(&self) -> bool

Test if this node has no subnodes or attributes

fn has_node(&self, name: &String) -> bool

Test if this node has a subnode with the given name.

fn has_nodes(&self) -> bool

Test if this node has any subnodes at all.

fn node_count(&self) -> usize

Returns the number of subnodes.

fn has_attr(&self, name: &String) -> bool

Test if this node had an attribute with the given key.

fn has_attrs(&self) -> bool

Test if this node has any attributes at all.

fn attr_count(&self) -> usize

Returns the number of attributes.